DiffSize bug fix

+++ include/maidsafe/nfs/vault/messages.h
@@ -68,7 +68,7 @@ void swap(AvailableSize& lhs, AvailableSize& rhs) MAIDSAFE_NOEXCEPT;
 // ================================ DiffSize =======================================================
 
 struct DiffSize {
-  explicit DiffSize(uint64_t size);
+  explicit DiffSize(int64_t size);
   DiffSize(const DiffSize& other);
   DiffSize(DiffSize&& other);
   DiffSize& operator=(DiffSize other);
@@ -76,7 +76,7 @@ struct DiffSize {
   explicit DiffSize(const std::string& serialised_copy);
   std::string Serialise() const;
 
-  int32_t diff_size;
+  int64_t diff_size;
 };
 
 bool operator==(const DiffSize& lhs, const DiffSize& rhs);
